+++ src/playground/configs/templates/ClientTable.vue
@@ -29,6 +29,7 @@
 			:level="tableSettings.level"
 			:levels="tableSettings.levels"
 			:loader-height="tableSettings.loaderHeight"
+			:loader-props="tableSettings.loaderProps"
 			:loader-size="tableSettings.loaderSize"
 			:loader-type="tableSettings.loaderType"
 			:loading="tableSettings.loading"
@@ -41,8 +42,11 @@
 			:page-text="tableSettings.pageText"
 			:prev-icon="tableSettings.prevIcon"
 			:search="tableSettings.search"
+			:search-container-cols="tableSettings.searchContainerCols"
 			:search-debounce="tableSettings.searchDebounce"
+			:search-events="tableSettings.searchEvents"
 			:search-max-wait="tableSettings.searchMaxWait"
+			:search-props="tableSettings.searchProps"
 			:select-strategy="tableSettings.selectStrategy"
 			:separator="tableSettings.separator"
 			:server="tableSettings.server"
@@ -70,7 +74,7 @@
 				[top Slot]
 			</template> -->
 
-			<!-- <template #[`top.left`]="{ props }">
+			<!-- <template #[`top.left`]="props">
 				<v-col
 					v-if="props.level === 1"
 					cols="4"