+++ src/types/index.ts
@@ -3,8 +3,8 @@
 import { CSSProperties, JSXComponent, StyleValue, MaybeRef } from 'vue';
 import { IconOptions, ThemeInstance } from 'vuetify';
 import type { EventBusKey } from '@vueuse/core';
-import type { VTextField, VProgressCircular, VProgressLinear } from 'vuetify/components';
-import type { VDataTable, VDataTableServer, VDataTableRow } from 'vuetify/labs/components';
+import type { VProgressCircular, VProgressLinear } from 'vuetify/components';
+import type { VDataTable, VDataTableServer, VDataTableRow, VSkeletonLoader } from 'vuetify/labs/components';
 
 
 // -------------------------------------------------- Vuetify Types //
@@ -46,6 +46,12 @@ export interface DataTableItem<T = any> {
 }
 
 
+// -------------------------------------------------- Misc //
+export interface KeyStringAny<T = any> {
+	[key: string]: T;
+};
+
+
 // -------------------------------------------------- Colors //
 export interface TableColors<T = any> {
 	[key: string]: T;
@@ -71,13 +77,6 @@ export type ColorsObject = {
 		bg?: string;
 		text?: string;
 	};
-	loader?: {
-		bg?: string;
-		circular?: string;
-		color?: string;
-		linear?: string;
-		text?: string;
-	};
 	percentageChange?: number;
 	percentageDirection?: 'asc' | 'desc';
 };
@@ -97,7 +96,7 @@ export type RGBColor = [number, number, number];
 
 
 // -------------------------------------------------- Search //
-export type SearchPropCols = {
+export type SearchContainerCols = {
 	lg?: number;
 	md?: number;
 	sm?: number;
@@ -106,28 +105,22 @@ export type SearchPropCols = {
 	xxl?: number;
 };
 
-export type SearchProps = {
-	cols?: SearchPropCols;
-	density?: VTextField['$props']['density'];
-	variant?: VTextField['$props']['variant'];
-};
-
 
 // -------------------------------------------------- Cell Rendering //
 interface CellRender {
 	(
-		key?: string,
 		column?: object,
 		index?: number,
+		key?: string,
 	): void;
 };
 
 interface ItemCellRender {
 	(
-		itemValue?: string,
-		item?: object,
 		column?: object,
 		index?: number,
+		item?: object,
+		itemValue?: string,
 	): void;
 };
 
@@ -190,7 +183,7 @@ export interface Props {
 	itemsPerPageOptions?: VDataTable['$props']['itemsPerPageOptions'];
 	level: number;
 	levels: number;
-	loaderHeight?: VProgressLinear['$props']['height'];
+	loaderProps?: LoaderProps;
 	loaderSize?: VProgressCircular['$props']['size'];
 	loaderType?: string | string[] | false | null;
 	loading?: VDataTable['$props']['loading'];
@@ -204,9 +197,11 @@ export interface Props {
 	page?: VDataTable['$props']['page'];
 	returnObject?: VDataTable['$props']['returnObject'];
 	search?: string | undefined;
+	searchContainerCols?: SearchContainerCols;
 	searchDebounce?: number | undefined | null;
+	searchEvents?: KeyStringAny;
 	searchMaxWait?: number | undefined | null;
-	searchProps?: SearchProps;
+	searchProps?: KeyStringAny;
 	separator?: 'default' | 'horizontal' | 'vertical' | 'cell' | undefined;
 	server?: boolean;
 	selectStrategy?: VDataTable['$props']['selectStrategy'];
@@ -215,7 +210,6 @@ export interface Props {
 	showFooterRow?: boolean;
 	showSearch?: boolean;
 	showSelect?: VDataTable['$props']['showSelect'];
-	skeltonType?: string;
 	sortAscIcon?: VDataTable['$props']['sortAscIcon'];
 	sortBy?: VDataTable['$props']['sortBy'];
 	tableType?: TableType;
@@ -266,7 +260,9 @@ export interface TopSlotProps extends VDataTableSlotProps {
 	items: Props['items'];
 	level: Props['level'];
 	loading: Props['loading'];
-	searchProps?: SearchProps;
+	searchContainerCols: SearchContainerCols;
+	searchEvents?: KeyStringAny;
+	searchProps?: KeyStringAny;
 	showSearch: boolean;
 };
 
@@ -274,6 +270,7 @@ export interface HeaderSlotProps extends AllSlotProps {
 	columnWidths: Props['columnWidths'];
 	isTheadSlot?: boolean;
 	items: Props['items'];
+	loaderProps: LoaderProps;
 	loaderSettings: {
 		colspan: number;
 		height?: VProgressLinear['$props']['height'];
@@ -281,7 +278,6 @@ export interface HeaderSlotProps extends AllSlotProps {
 		loading: VDataTable['$props']['loading'];
 		loadingText?: VDataTable['$props']['loadingText'];
 		size?: VProgressCircular['$props']['size'];
-		skeltonType: Props['skeltonType'];
 		textLoader?: boolean;
 	};
 	matchColumnWidths: Props['matchColumnWidths'];
@@ -357,6 +353,29 @@ export interface TFootSlotProps extends Omit<AllSlotProps, 'showSelect' | 'sortB
 
 
 // -------------------------------------------------- Table Loader //
+export interface CircularLoaderProps extends KeyStringAny {
+	bgColor?: VProgressCircular['$props']['bgColor'];
+	color?: VProgressCircular['$props']['color'];
+	indeterminate?: VProgressCircular['$props']['indeterminate'];
+	size?: VProgressCircular['$props']['size'];
+}
+export interface LinearLoaderProps extends KeyStringAny {
+	color?: VProgressLinear['$props']['color'];
+	indeterminate?: VProgressLinear['$props']['indeterminate'];
+}
+export interface SkeltonLoaderProps extends KeyStringAny {
+	type?: VSkeletonLoader['$props']['type'],
+}
+
+export interface LoaderProps {
+	circular?: CircularLoaderProps,
+	linear?: LinearLoaderProps,
+	skelton?: SkeltonLoaderProps,
+	text?: {
+		color?: string,
+	};
+};
+
 export type TableLoader = {
 	colors: Props['colors'];
 	colspan: number;
@@ -364,9 +383,9 @@ export type TableLoader = {
 	level: Props['level'];
 	loaderType: Props['loaderType'];
 	loading: VDataTable['$props']['loading'];
+	loaderProps: LoaderProps;
 	loadingText?: VDataTable['$props']['loadingText'];
 	size?: VProgressCircular['$props']['size'];
-	skeltonType: Props['skeltonType'];
 	textLoader?: boolean;
 };
 
@@ -374,7 +393,7 @@ export interface UseLoaderStyles {
 	(
 		options: {
 			isLinearOnly: MaybeRef<boolean>,
-			loaderHeight: MaybeRef<Props['loaderHeight']>,
+			loaderHeight: MaybeRef<VProgressLinear['$props']['height']>,
 		}
 	): CSSProperties;
 }